frontend: remove old dashboard.
